% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{c.lisst}
\alias{c.lisst}
\title{Concatenate lisst objects}
\usage{
\method{c}{lisst}(...)
}
\value{
A lisst object with type equal to the first argument and containing 
data from all arguments.
}
\description{
This function allows to concatenate two or more lisst objects into a single
object. Checks are made to ensure the final object is coherent (same 
instrument and inversion type, if applicable) and conversion between types is
made as necessary. See details.
}
\details{
The first lisst object is taken as reference, i.e., all subsequent 
lisst objects will be converted to its type ('raw', 'cor', 'cal', 'vsf', 
'vol' or 'pnc') before the concatenation. A check on the lisst attributes is 
made to ensure consistency that all data is from a single instrument/model 
and from the same inversion type (if applicable).
}
\examples{
l1 <- donkmeer_bin
l2 <- lget(donkmeer_bin, 'raw')
l3 <- c(l1, l2)

}
