% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{.getmodp}
\alias{.getmodp}
\title{For the LISST-100 series, extract only the relevant model information for the 
detector type.}
\usage{
.getmodp(linst)
}
\description{
For the LISST-100 series, extract only the relevant model information for the 
detector type.
}
