% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\docType{data}
\name{.LISSTm}
\alias{.LISSTm}
\title{.LISSTm - A list of model dependent parameters list for each supported model.}
\format{An object of class \code{list} of length 2.}
\usage{
.LISSTm
}
\description{
Variables:
pl    - Standard path length (m) of the instrument.
bnvar - Number of variables in the binary file.
anvar - Number of variables in the processed file.
nring - Number of detectors.
a0    - Minimum angle in air for the inner most ring (rad).
s0    - Minimum size range (µm) per inversion type (ss or rs) and per 
detector type (model dependent).
lvarn - Long variable names.
varun - Variable units.
aang  - In air angles (rad) for each detector ring (assinged to the correct 
bins), per detector type (model dependent).
wang  - In water angles (rad) for each detector ring(assinged to the correct 
bins), per detector type (model dependent).
binr  - Bin size range (µm) per inversion type (ss or rs) and per detector 
type (model dependent).
}
\details{
Note: Informations were retrieved from LISST user manuals for the respective
models, except the water angles for the LISST-200X, that were taken from the
LISST-VSF manual since the ring detector assembly is the same as in the 
LISST-200X, while the LISST-200X manual don't provide angle information.
}
\keyword{datasets}
