% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\name{drop_lisst}
\alias{drop_lisst}
\title{drop lisst}
\usage{
drop_lisst(x)
}
\arguments{
\item{x}{A lisst object.}
}
\value{
A \code{data.frame}, with all columns of class \code{quantities} in 
exception of those that contain time information.
}
\description{
Drop lisst class and attributes.
}
\examples{
x <- drop_lisst(donkmeer_bin)
attributes(x)

}
