% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fprofile}
\alias{fprofile}
\title{Fictional data of LISST measurements in a profiling system}
\format{A 'vol' lisst object with 41 records.}
\usage{
fprofile
}
\description{
This dataset contains fictional data simulating a measurement setup in a 
profiling platform, profiling a 10 m deep water.
}
\keyword{datasets}
