% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getvsf}
\alias{getvsf}
\title{Retrieve the Particle Volume Scattering Function (VSF) from calibrated LISST 
data.}
\usage{
getvsf(lcal)
}
\arguments{
\item{lcal}{object of class lisst and type cal.}
}
\description{
Retrieve the Particle Volume Scattering Function (VSF) from calibrated LISST 
data.
}
\references{
Agrawal, Yogesh C. 2005. The optical volume scattering function: Temporal and 
vertical variability in the water column off the New Jersey coast. Limnology 
and Oceanography 50, 6, 1787-1794. DOI: 10.4319/lo.2005.50.6.1787
}
