% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\name{is.lisst}
\alias{is.lisst}
\title{Test for a lisst object}
\usage{
is.lisst(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A logical value indicating if the argument is a lisst object.
}
\description{
Check that an object loosely conforms to a lisst object.
}
\examples{
is.lisst(donkmeer_pro)
is(donkmeer_pro, 'lisst')

}
