% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\name{lbinr}
\alias{lbinr}
\title{Extract size range of bins}
\usage{
lbinr(x, ity)
}
\arguments{
\item{x}{A lisst object.}

\item{ity}{Inversion type ('ss' or 'rs').}

\item{mod}{A LISST model ('100(X)B', '100(X)C' or '200X').}
}
\value{
A \code{units} \code{matrix} with lower and upper size limits and 
median size in µm.
}
\description{
Extracts the size range and median size per bin per LISST model and particle
inversion type.
}
\details{
If x is provided, mod and ity (if present) will be extracted from the 
object's metadata. If provided, mod and ity will take precedence. Note that
since inversion is not implemented in this version, only lisst objects 
created from LISST SOP processed files will contain ity information.
}
\examples{
lbinr(mooring)
lbinr(,'100XC', 'rs')

}
