% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lfit}
\alias{lfit}
\title{Fit a PSD model to the lisst data}
\usage{
lfit(x, type)
}
\arguments{
\item{x}{A PSD ('vol' or 'pnc') lisst object.}

\item{model}{Curently ignored. Only Junge model implemented.}
}
\description{
The function performs a fit of the available models of PSD.
}
