% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgdate}
\alias{lgdate}
\title{Get LISST measurement dates}
\usage{
lgdate(lo, yr)
}
\arguments{
\item{lo}{A lisst object of type 'raw', 'cor' or 'cal'.}

\item{yr}{The year of first measurement in the lisst object. Can be omitted. 
Ignored for LISST-200X.}
}
\description{
The function retrieves the date/time of measurements for the records in the 
lisst object as POSIXct objects. The time zone of the dates will depend on 
user tz options.
}
\details{
If yr is not provided (or is NULL) teh function will perform a 'best 
guess' (with a warning) based on the curent date and the julian date of first 
measurement. Simply, if the julian day of first measurement is greater than
the curent julian day, the previous from current year is assumed and the 
current year otherwise.
}
