% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lget}
\alias{lget}
\title{Convert between LISST data representations}
\usage{
lget(x, type)
}
\arguments{
\item{x}{A lisst object.}

\item{type}{A character specifing the lisst data type. One of 'raw', 'cor', 
'cal', 'vsf', 'vol', 'pnc'. See details.}
}
\description{
This set of functions allow for the convertion between lisst data types.
}
\details{
lget dispatch the appropriate lgetxxx function acording to parameter 
type. As of this versions is only possible to convert between 'raw', 'cor', 
'cal' and 'vsf' or between 'vol' and 'pnc' (inversion not implemented). 
\describe{
  \item{raw}{The raw digital counts as recorded by the LISST instrument}
  \item{cor}{The corrected digital counts, i.e., the raw counts de-attenuated 
             for the particle \strong{and water} extinction, background 
             subtracted and compensated for area deviations from nominal 
             values}
  \item{cal}{The calibrated values, i.e., the instrument specific calibration 
             constants applied to the corrected values (for all variables). 
             Aditionally, the transmittance due to particles and the particle 
             beam attenuation are added to the lisst object}
  \item{vsf}{The volume scattering function, i.e., the calibrated values 
             normalized to incident power, the solid angle of the detectors 
             and the path of water generating the signal}
  \item{vol}{The particle volume concentration (ppm volume) per size bin, as 
	inverted from the scattering data by the LISST-SOP}
  \item{pnc}{The particle number concentration per volume and µm size. }
}

See the documentation of the lget functions for further details.
}
\seealso{
\code{\link{lgetraw}}, \code{\link{lgetcor}}, \code{\link{lgetval}},
\code{\link{lgetvsf}}, \code{\link{lgetvol}}, \code{\link{lgetpnc}}
}
