% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgetcal}
\alias{lgetcal}
\title{Retrieve the LISST calibrated values}
\usage{
lgetcal(x)
}
\arguments{
\item{x}{A lisst object of type 'raw', 'cor', 'cal' or 'vsf'.}
}
\description{
The function retrieve the calibrated values (lisst object type 'cal') from 
raw digital counts ('raw'), corrected digital counts ('cor') or volume 
scattering function ('vsf').
}
\details{
When supplying a lisst 'raw' object, it must contain information on 
background values and instrument specific calibration constants. This is the
case when the 'raw' lisst object was created with zscat and sn arguments set
through \code{read_lisst}.
}
\examples{
l_cal <- lgetcal(donkmeer_bin)

}
\seealso{
\code{\link{lget}} \code{\link{lgetraw}}, \code{\link{lgetcor}}, 
\code{\link{lgetvsf}}, \code{\link{lgetvol}}, \code{\link{lgetpnc}},
\code{\link{lgetcsa}}
}
