% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgetcor}
\alias{lgetcor}
\title{Retrieve the LISST corrected digital counts}
\usage{
lgetcor(lo)
}
\arguments{
\item{lo}{A lisst object of type 'raw', 'cor' or 'cal'.}
}
\description{
The function retrieve the corrected digital counts (lisst object type 'cor')
from raw digital counts ('raw') or from calibrated values ('cal').
}
\details{
When supplying a lisst 'raw' object, it must contain information on 
background values and instrument specific calibration constants. This is the
case when the 'raw' lisst object was created with zscat and sn arguments set
through \code{read_lisst}.

For consistency with LISST processed data provided by manufacturer, the 
optical transmission and the beam attenuation do not include the effect of 
pure water. However, the digital counts in the ring detectors are corrected 
for the aditional transmission loss due to pure water absorption and 
scattering. Values for the pure water absorption and scattering at 670 nm are 
taken from the Water Optical Properties Processor (WOPP) and correspond to 
0.439 and 5.808e-4 1/m, respectivelly.
}
