% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgetcsa}
\alias{lgetcsa}
\title{Retrieve PSD in cross sectional area concentration}
\usage{
lgetcsa(x)
}
\arguments{
\item{x}{A lisst object of type 'vol' or 'pnc'.}
}
\description{
The function converts the PSD in volume concentration (µL/L, ppm) to cross 
sectional concentration (m2/m3).
}
\details{
Volume concentration is converted to cross sectional concentration by 
using the cross section and volume of a sphere with radius equal to half the 
median particle size for each bin. The cross sectional concentration is then the 
sphere equivalent cross section concentration.
}
\examples{
l_csa <- lgetcsa(donkmeer_pro)

}
\references{
Slade, W. H. and Boss, E. S. 2015. Spectral attenuation and backscattering as 
indicators of average particle size. Applied Optics 54, 24, 7264-7277. 
DOI: 10.1364/AO.54.007264
}
\seealso{
\code{\link{lget}} \code{\link{lgetraw}}, \code{\link{lgetcor}}, 
\code{\link{lgetcal}}, \code{\link{lgetvsf}}, \code{\link{lgetvol}},
\code{\link{lgetpnc}}
}
