% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lgetfit}
\alias{lgetfit}
\title{Fit a PSD model to the lisst data}
\usage{
lgetfit(lo, type)
}
\arguments{
\item{lo}{A PSD ('vol' or 'pnc') lisst object.}

\item{type}{Curently ignored.}
}
\description{
The function performs a fit of the available models of PSD.
}
