% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgetraw}
\alias{lgetraw}
\title{Retrieve the LISST raw digital counts}
\usage{
lgetraw(x)
}
\arguments{
\item{x}{A lisst object of type 'raw', 'cor', 'cal' or 'vsf'.}
}
\description{
The function retrieve the raw digital counts (lisst object type 'raw')
from corrected digital counts ('cor'), calibrated values ('cal'), or  
volume scattering function ('vsf').
}
\seealso{
\code{\link{lgetcor}}, \code{\link{lgetval}}, \code{\link{lgetvsf}}, 
\code{\link{lgetvol}}, \code{\link{lgetpnc}}
}
