% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lhov}
\alias{lhov}
\title{Hovmöller diagram for lisst objects}
\usage{
lhov(x, by = "sample", nbins = "pretty", norm = TRUE, legend = TRUE,
  col, xlab, ylab, zlab, yu, zu)
}
\arguments{
\item{x}{A lisst object.}

\item{by}{Ordinate axis dimension. Must be one of 'sample', 'time' or 
'depth'.}

\item{nbins}{If the lisst object is not regular in the dimension specified 
parameter by, nbins controls the binning of the data to achieve 
a regular series appropriate for raster representation.}

\item{norm}{Logic. Should the magnitude values de normalized? See details.}

\item{legend}{Logic. Should a legend be added to the plot?}

\item{col}{A color pallete to be used. If not specified a default pallete 
will be used.}

\item{yu}{The units for the abscissa. See details.}

\item{zu}{The units for the magnitude values. See details.}
}
\description{
Create Hovmöller diagram for lisst objects based on sample number, 
measurement time or depth.
}
\details{
The z scale is provided in log10 for appropriate color mapping. However, 
LISST 100 SOP will return x volume concentration if concentration is below 
0.001 ppm. In that case, zeros as treated as half this minimul value, 0.0005 
ppm.

If the lisst object is not regularly spaced in the chosen dimension, it will 
be binned for raster representation. The nbins parameter will control the 
number of bins, with all bins having the same interval size. The sample 
dimension will always be regular.

Normalization is performed by the sum of all detector values. It is provided 
to aid visual separation from changes in concentration and distribution. If 
TRUE a top pannel is added to the plot, with a line plot indicating the 
magnitude of the sum of detector values.

If legend is set to TRUE a bottom pannel will be added to the plot with a 
magnitude scale. The axis label and units are affected by the zlab and zu 
parameters.

Note that unit conversion is performed 'automatcaly' using the units package. 
If there is no standard conversion between the default and the request units
the function will exit with an error. See \code{link{units}} for details.
}
\examples{
flp <- system.file("extdata", "DN_27_rs.asc", package = "lisst")
model <- "100CX" 
lop <- read_lisst(flp, model = model)
lhov(lop, by = 'sample')
lhov(lop, by = 'sample', norm = FALSE)
lhov(lop, by = 'sample', norm = FALSE, legend = FALSE)

}
