% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{linst}
\alias{linst}
\title{Retrieve a list of registered instruments}
\usage{
linst()
}
\value{
A character vector.
}
\description{
The function provide a character vector of the installed instruments 
contaning model and serial number.
}
