% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\docType{package}
\name{lisst-package}
\alias{lisst-package}
\title{\pkg{lisst}: Utilities for analysis of LISST data}
\description{
An R package to read, manipulate and visualize data from the Laser In-Situ 
Scattering and Transmissometry (LISST) instruments, Sequoia Scientific, Inc.
}
\author{
Alexandre Castagna
}
