% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{lisst_bin}
\alias{lisst_bin}
\title{Read a LISST binary file}
\usage{
lisst_bin(fl, zscat, sn, pl, out)
}
\description{
Read a LISST binary file for the registered instruments of supported models. 
It is not intended to be used directly, but called from \code{read_lisst}.
}
