% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{lisst_pro}
\alias{lisst_pro}
\title{Read a LISST processed file}
\usage{
lisst_pro(fl, sn, pl, zscat, linst, lmodl, yr)
}
\description{
Read a LISST processed file for the registered instruments of supported 
models. It is not intended to be used directly, but called from 
\code{read_lisst}.
}
