% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{lisst_reg}
\alias{lisst_reg}
\title{Register a new LISST instrument}
\usage{
lisst_reg(model, path)
}
\arguments{
\item{model}{The model of the LISST instrument (e.g., "100" for the LISST-100
and LISST-100X).}

\item{path}{Path to directory containing instrument instalation files. See 
details.}
}
\description{
Register a new LISST instrument by importing calibration data and storing it 
for convenient processing of LISST data.
}
\details{
For the LISST-100(X) the function expects that files follow the 
naming convention for LISST installation files. Specifically, it will search 
for a factory background file (factory_zsc_*.asc), a ring area file 
(ringarea_*.asc), an instrument data file (InstrumentData.txt) and a 
LISST-SOP ini file (lisst.ini). All files should be in the same directory 
provided in the path argument, and it should contain files for a single 
instrument.
}
\examples{
:
path  <- system.file("extdata", package = "lisst")
model <- 100
lisst_reg(model, path)

}
