% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lrings}
\alias{lrings}
\title{Sum of ring values}
\usage{
lrings(x, rings)
}
\arguments{
\item{x}{A lisst object.}

\item{rings}{The rings to be summed.}
}
\description{
The function performs a sum over the specified rings.
}
\details{
If parameter rings is not supplied, sum will be performed over all rings.
}
