% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\name{lwang}
\alias{lwang}
\title{Extract angle range of bins}
\usage{
lwang(x)
}
\arguments{
\item{x}{A lisst object or a LISST model ('100(X)B', '100(X)C' or '200X').}
}
\value{
A \code{units} \code{matrix} with lower and upper angle limits and 
median angle in radians.
}
\description{
Extracts the (in water) angle range and median angle per bin per LISST model.
}
\examples{
lwang(mooring)
lwang('100XC')

}
