% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{metadata}
\alias{metadata}
\alias{zscat}
\alias{fzscat}
\alias{lity}
\alias{ltime}
\alias{angles}
\alias{sizes}
\title{Extract metadata from lisst objects}
\usage{
zscat(x)

fzscat(x)

lity(x)

ltime(x)

angles(x, units, medium = c("water", "air"))

sizes(x, units)
}
\arguments{
\item{x}{A lisst object.}

\item{units}{Desired units.}

\item{medium}{Desired medium (water or air).}
}
\value{
The returned data is dependent on the specific metadata requested:

\itemize{
  \item ltyi   - A character indicating the LISST SOP inversion type (rs or ss) if any;
  \item sizes  - A numeric matrix with the minimum, maximum and median sizes in each bin;
  \item angles - A numeric matrix with the minimum, maximum and median angles in each bin;
  \item ltime  - A POSIXct vector of recording times;
  \item zscat  - A numeric vector with background scattering values (blank);
  \item fzscat - A numeric vector with factory background scattering values (blank).
}
}
\description{
This collection of functions allows to extract specific metadata stored in the lisst
object.
}
\details{
These functions are the recommended way to access the object's metadata since 
metadata organization might be subject to change.
}
\section{Functions}{
\itemize{
\item \code{zscat}: Extract background scattering

\item \code{fzscat}: Extract factory background scattering

\item \code{lity}: Extract inversion type

\item \code{ltime}: Extract lisst time index

\item \code{angles}: Extract reference measuring angles

\item \code{sizes}: Extract reference inversion sizes
}}

