% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mooring}
\alias{mooring}
\title{Mooring of LISST-200X}
\format{An object of class \code{lisst} (inherits from \code{data.frame}) with 782 rows and 61 columns.}
\usage{
mooring
}
\description{
Mooring of LISST-200X
}
\keyword{datasets}
