% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBins}
\alias{plotBins}
\title{Plot Bins}
\usage{
plotBins(x, bins, by, col, lty)
}
\arguments{
\item{x}{A lisst object.}

\item{bins}{A list of the rings to be summed. See details and examples.}

\item{by}{Ordinate. One of 'sample', 'time' or 'depth'}

\item{col}{A vector of colors.}

\item{lty}{A vector of line types.}
}
\description{
Plot Bins
}
