% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_lisst}
\alias{read_lisst}
\title{Read LISST data}
\usage{
read_lisst(fl, zscat, sn, pl, yr, out)
}
\arguments{
\item{fl}{Path to binary file (e.g., *.DAT in the LISST-100(X)).}

\item{zscat}{Path to ASCII background data file for LISST-100(X) (e.g., 
*.asc). Ignored for LISST-200X.}

\item{sn}{Serial number of the instrument.}

\item{pl}{Path length in meters. If not provided the system will assumed 
standard path length for the instrument model, i.e., no path 
reduction module in use.}

\item{yr}{The year of first measurement in the data file. Only necessary 
for LISST-100(X).}

\item{out}{The output format. Valid options are 'vol', 'psd', 'raw', 'cor' 
and 'cal'. See details.}
}
\value{
A lisst object with attribute type according to the processing level 
requested.
}
\description{
Read LISST data (processed or binary files) for the registered instruments of
supported models.
}
\details{
The function will determine the file type based on its extenssion. 
Processed files outputed from LISST SOP for the LISST-100(X) have extension 
.asc and from LISST-200X, extension .csv. Binary files have extension .DAT 
and RBN for LISST-100(X) and LISST-200X, respectivelly.

The level of processing will depend on the output requested by the 
user. For processed LISST files as outputed by the LISST SOP the volume 
concentration ('vol', ppm) or the particle number concentration ('pnc', 
1/L/µm) can be returned. For binary files, the raw digital counts ('raw'), 
the corrected digital counts ('cor') or the calibrated values ('cal') can be
returned. Corrected digital counts are the digital counts of the ring 
detectors de-attenuated, background corrected and compensated for ring area 
deviations from ideal log increase behaivour. All other parameters are in 
'cor' are kept at digital counts. Finally, for 'cal', all parameters are 
converted to physical units using the calibration constants registered with 
\code{lisst_reg} for the instrument serial number.
}
