% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{lgetpnc}
\alias{lgetpnc}
\title{Retrieve PSD in number concentration}
\usage{
lgetpnc(x)
}
\arguments{
\item{x}{A lisst object of type 'vol' or 'csa'.}
}
\description{
The function converts the PSD in volume concentration (µL/L, ppm) to number 
concentration (particle/L/µm).
}
\details{
Volume concentration is converted to number concentration by using 
the volume of a sphere with radius equal to half the median particle size
for each bin. The number concentration is then the sphere equivalent number
concentration. The absolute magnitute of the PSD will be only approximate if 
particles are not spherical, but as long as the particles are not expected to
significantly change shape with size, the slope of the distribution will be 
accurate.
}
\examples{
l_pnc <- lgetpnc(donkmeer_pro)

}
\references{
Buonassissi, C. J. and Dierssen, H. M. 2010. A regional comparison of 
particle size distributions and the power law approximation in oceanic and 
estuarine surface waters. J. Geophys. Res., 115, C10028. 
DOI:10.1029/2010JC006256.
}
\seealso{
\code{\link{lget}} \code{\link{lgetraw}}, \code{\link{lgetcor}}, 
\code{\link{lgetcal}}, \code{\link{lgetvsf}}, \code{\link{lgetvol}},
\code{\link{lgetcsa}}
}
