% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lstat}
\alias{lstat}
\alias{mean.lisst}
\alias{median.lisst}
\title{Descriptive statistics for lisst objects}
\usage{
lstat(x, brks, fun = "mean", ...)

\method{mean}{lisst}(x, ...)

\method{median}{lisst}(x, ...)
}
\arguments{
\item{x}{A lisst object.}

\item{brks}{A vector with the breaks (intervals) for the aggregation. See 
details.}

\item{fun}{A function to perform the aggregation (mean or median). Defaults 
to mean.}

\item{...}{Arguments to be passed to aggregation functions.}
}
\description{
The functions calculate the average or the median per variable of a lisst 
object, for the whole data or in intervals. The dispersion of the dataset or
intervals is also retrieved and stored.
}
\details{
The breaks (intervals) are passed directly to the subset function, so must 
now be supplied in final form.

The measure of dispersion in the aggregated data depends on the measure of 
center. For the mean the dispersion is the standard deviation and for the 
median it is 1.252 * the standard deviation of the mean, as derived from the 
asymptotic variance formula of the median. See \code{?mean.errors} for 
details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{mean}: Compute the mean for lisst objects

\item \code{median}: Compute the median for lisst objects
}}

\examples{


}
