% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_lisst}
\alias{read_lisst}
\title{Read LISST data}
\usage{
read_lisst(fl, sn, pl, zscat, yr, out, model)
}
\arguments{
\item{fl}{Path to processed or binary file (e.g., *.DAT in the 
LISST-100(X)).}

\item{sn}{Serial number of the instrument. Optionally, can be omitted when 
reading a SOP processed file or when out is 'raw' or 'cor' for 
binary files. In this case, instrument model must be supplied to 
argument model.}

\item{pl}{Path length in meters. If not provided the function will assume 
standard path length for the instrument model (with a warning), 
i.e., no path reduction module in use.}

\item{zscat}{Path to ASCII background data file for LISST-100(X) (e.g., 
*.asc). Ignored for LISST-200X. Only required when reading 
binary files.}

\item{yr}{The year of first measurement in the data file. Only necessary 
for LISST-100(X). If not provided, function will make a best 
guess (with a warning) based on the curent date and the julian
date of first measurement.}

\item{out}{The output format. Valid options are 'vol', 'pnc', 'raw', 'cor' 
and 'cal'. See details.}

\item{model}{A character vector of the instrument model (e.g. "200" for 
LISST-200X). For the LISST-100(X), the detector type must be 
included in the name (e.g., "100C" or "100CX"). Ignored if sn is 
provided.}
}
\value{
A lisst object with attribute type according to the processing level 
requested.
}
\description{
Read LISST data (processed or binary files) for the registered instruments of
supported models.
}
\details{
The function will determine the file type based on its extenssion. 
Processed files created from LISST SOP for the LISST-100(X) have extension 
.asc and from LISST-200X, extension .csv. Binary files have extension .DAT 
and RBN for LISST-100(X) and LISST-200X, respectivelly. The extension is also
used to determine 

The level of processing will depend on the output requested by the 
user. For processed LISST files as creayed by the LISST SOP, the volume 
concentration ('vol', ppm) or the particle number concentration ('pnc', 
1/L/µm) can be returned. For binary files, the raw digital counts ('raw'), 
the corrected digital counts ('cor') or the calibrated values ('cal') can be
returned. Corrected digital counts are the digital counts of the ring 
detectors de-attenuated, background corrected and compensated for ring area 
deviations from ideal log increase behaivour. All other parameters in 'cor' 
are kept at the original digital counts. Finally, for 'cal', all parameters 
are converted to physical units using the calibration constants registered 
with \code{lisst_reg} for the instrument serial number. In addition, types 
'cal','vol' and 'pnc' return an extra column, with dates/times converted to 
POSIXct format.

To allow easier manipulation, conversion between types, and plotting, all 
associated information for the instrument and model are saved as attributes 
in the returned object.
}
\references{
MATLAB source code provided by Sequoia Scientific, Inc, and available at:
https://www.sequoiasci.com/product/lisst-100x/, 
https://www.sequoiasci.com/product/lisst-200x/
}
